+++ esp-hal/src/soc/esp32s2/ulp_core.rs
@@ -48,33 +48,33 @@ use crate::peripherals::LPWR;
 pub enum UlpCoreWakeupSource {
     /// Wakeup source from the HP (High Performance) CPU.
     HpCpu,
+    /// Wakeup after the ULP Timer has elapsed.
+    /// The actual period between wake-ups is affected by the runtime duration of the ULP program.
+    Timer(UlpCoreTimerCycles),
 }
 
-/// Configures `RTC_CNTL_ULP_CP_TIMER_1_REG` to set sleep cycles for ULP coprocessor timer.
+/// ULP Timer cycles are clocked at a rate of approximately 8MHz / 32768  = ~244 Hz.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct UlpCoreSleepCycles {
+pub struct UlpCoreTimerCycles {
     cycles: u32,
 }
-impl UlpCoreSleepCycles {
-    /// Creates a new sleep cycle configuration.
+impl UlpCoreTimerCycles {
+    /// Creates a new Ulp Timer cycle count configuration.
     /// ## Panics
     ///
     /// Panics if the cycles value is outside of the value range (0 ..= 0xFFFFFF).
     pub const fn new(cycles: u32) -> Self {
         ::core::assert!(
             cycles <= 0xFFFFFF,
-            "`RTC_CNTL_ULP_CP_TIMER_SLP_CYCLE` sleep cycles must be between 0 and 0xFFFFFF (inclusive)."
+            "ULP Timer cycles must be between 0 and 0xFFFFFF (inclusive)."
         );
         Self { cycles }
     }
     fn cycles(self) -> u32 {
         self.cycles
     }
-    fn value(self) -> u32 {
-        self.cycles()
-    }
 }
-impl Default for UlpCoreSleepCycles {
+impl Default for UlpCoreTimerCycles {
     fn default() -> Self {
         // ESP32-S2 Technical Reference Manual. Register 1.2. RTC_CNTL_ULP_CP_TIMER_1_REG (0x0130)
         // Field RTC_CNTL_ULP_CP_TIMER_SLP_CYCLE has a default value of 200 cycles.
@@ -85,7 +85,6 @@ impl Default for UlpCoreSleepCycles {
 /// Structure representing the ULP (Ultra-Low Power) core.
 pub struct UlpCore<'d> {
     _lp_core: crate::peripherals::ULP_RISCV_CORE<'d>,
-    _sleep_cycles: UlpCoreSleepCycles,
 }
 
 impl<'d> UlpCore<'d> {
@@ -96,19 +95,7 @@ impl<'d> UlpCore<'d> {
             unsafe { core::slice::from_raw_parts_mut(0x5000_0000 as *mut u32, 8 * 1024 / 4) };
         lp_ram.fill(0u32);
 
-        Self {
-            _lp_core: lp_core,
-            _sleep_cycles: Default::default(),
-        }
-    }
-
-    /// Sets the number of ULP Timer cycles to wait after the ULP halts, before starting it again.
-    /// ULP Timer cycles are clocked at a rate of approximately 8MHz / 32768  = ~244 Hz.
-    /// The actual period between wake-ups is affected by the runtime duration of the ULP program.
-    pub fn with_sleep_cycles(self, cycles: UlpCoreSleepCycles) -> Self {
-        let mut x = self;
-        x._sleep_cycles = cycles;
-        x
+        Self { _lp_core: lp_core }
     }
 
     // currently stopping the ULP doesn't work (while following the procedures
@@ -120,16 +107,8 @@ impl<'d> UlpCore<'d> {
 
     /// Runs the ULP core with the specified wakeup source.
     pub fn run(&mut self, wakeup_src: UlpCoreWakeupSource) {
-        ulp_set_wakeup_period(self._sleep_cycles);
         ulp_run(wakeup_src);
     }
-}
-
-fn ulp_set_wakeup_period(sleep_cycles: UlpCoreSleepCycles) {
-    let cycles = sleep_cycles.value() << 8;
-    LPWR::regs()
-        .ulp_cp_timer_1()
-        .write(|w| unsafe { w.ulp_cp_timer_slp_cycle().bits(cycles) });
 }
 
 #[allow(unused)] // TODO: remove cfg when implementation is corrected
@@ -193,7 +172,15 @@ fn ulp_run(wakeup_src: UlpCoreWakeupSource) {
 fn ulp_config_wakeup_source(wakeup_src: UlpCoreWakeupSource) {
     match wakeup_src {
         UlpCoreWakeupSource::HpCpu => {
-            // use timer to wake up
+            // only wake-up when the HpCpu calls .run()
+        }
+        UlpCoreWakeupSource::Timer(sleep_cycles) => {
+            // configure timer duration
+            let cycles = sleep_cycles.cycles() << 8;
+            LPWR::regs()
+                .ulp_cp_timer_1()
+                .write(|w| unsafe { w.ulp_cp_timer_slp_cycle().bits(cycles) });
+            // enable the timer
             LPWR::regs()
                 .ulp_cp_ctrl()
                 .modify(|_, w| w.ulp_cp_force_start_top().clear_bit());